REFACTOR CHECKPOINT: remove EF Core
EntityFramework Core is being removed from this project as it is really just
getting in the way of things. This commit is about rebuilding the project in
, which will then be moved as the root of this project once
the refactor is complete. Anything not in that directory is out of date and/or
un-processed.

This commit marks the migration of the basic entities (not their relations)
out of EF Core.