Partially undo the previous commit.
I can't just revert the whole commit because the previous commit is really two
commits. I want to keep the EntityData.Clone() code and the transaction code,
but I want to disregard all of the functionality that would allow a single
IRAMEntityRepo to Save() several different entities' datas. This is because
it was getting extremely cumbersome due to typing conflicts when I could just
take a different route.