Adjust ERD.
Vote was not correct, it contained the voting character, the praxis, and
the receiving character. It is now normalized to contain the voting
character and the praxis participant.

This involved tweaking Vote, IVoteRepo and RAMVoteRepo, and VoteDel.
Additionally, I needed to remove the logic from PraxisDel that dealt with
deleting the related Vo tes, and moving it to PraxisParticipantDel.